Core - user space separation
