+++ core/interfaces/interface-nbpc-admin-module.php
