+++ core/interfaces/interface-nbpc-module.php
