+++ core/interfaces/interface-nbpc-register.php
