+++ core/registers/class-nbpc-register-ajax.php
@@ -59,7 +59,7 @@ if ( ! class_exists( 'NBPC_Register_Ajax' ) ) {
 		}
 
 		public function get_items(): Generator {
-			yield null;
+			yield call_user_func( [ NBPC_Registers::class, 'regs_ajax' ], $this );
 		}
 	}
 }
