+++ core/registers/class-nbpc-register-style.php
@@ -1,6 +1,6 @@
 <?php
 /**
- * NBPC: Script register
+ * NBPC: Style register
  */
 
 /* ABSPATH check */
@@ -8,8 +8,8 @@ if ( ! defined( 'ABSPATH' ) ) {
 	exit;
 }
 
-if ( ! class_exists( 'NBPC_Register_Script' ) ) {
-	class NBPC_Register_Script implements NBPC_Register {
+if ( ! class_exists( 'NBPC_Register_Style' ) ) {
+	class NBPC_Register_Style implements NBPC_Register {
 		use NBPC_Hook_Impl;
 
 		public function __construct() {
@@ -22,17 +22,19 @@ if ( ! class_exists( 'NBPC_Register_Script' ) ) {
 		 */
 		public function register() {
 			foreach ( $this->get_items() as $item ) {
-				if ( $item instanceof NBPC_Reg_Script ) {
+				if ( $item instanceof NBPC_Reg_Style ) {
 					$item->register();
 				}
 			}
 		}
 
 		public function get_items(): Generator {
-			yield null;
+			yield call_user_func( [ NBPC_Registers::class, 'regs_style' ], $this );
 		}
 
 		/**
+		 * 'src' location helper.
+		 *
 		 * @param string $rel_path
 		 * @param bool   $replace_min
 		 *
@@ -41,11 +43,11 @@ if ( ! class_exists( 'NBPC_Register_Script' ) ) {
 		protected function src_helper( string $rel_path, bool $replace_min = true ): string {
 			$rel_path = trim( $rel_path, '\\/' );
 
-			if ( nbpc_script_debug() && $replace_min && substr( $rel_path, - 7 ) === '.min.js' ) {
-				$rel_path = substr( $rel_path, 0, strlen( $rel_path ) - 7 ) . '.js';
+			if ( nbpc_script_debug() && $replace_min && substr( $rel_path, - 8 ) === '.min.css' ) {
+				$rel_path = substr( $rel_path, 0, strlen( $rel_path ) - 8 ) . '.css';
 			}
 
-			return plugin_dir_url( nbpc()->get_main_file() ) . "assets/js/{$rel_path}";
+			return plugin_dir_url( nbpc()->get_main_file() ) . 'assets/css/' . $rel_path;
 		}
 	}
 }
