+++ core/registers/class-nbpc-register-submit.php
@@ -61,7 +61,7 @@ if ( ! class_exists( 'NBPC_Register_Submit' ) ) {
 		}
 
 		public function get_items(): Generator {
-			yield null;
+			yield call_user_func( [ NBPC_Registers::class, 'regs_submit' ], $this );
 		}
 	}
 }
