+++ core/registers/class-nbpc-register-taxonomy.php
@@ -29,7 +29,7 @@ if ( ! class_exists( 'NBPC_Register_Taxonomy' ) ) {
 		}
 
 		public function get_items(): Generator {
-			yield null;
+			yield call_user_func( [ NBPC_Registers::class, 'regs_taxonomy' ], $this );
 		}
 	}
 }
