+++ core/registers/class-nbpc-register-uninstall.php
@@ -26,7 +26,7 @@ if ( ! class_exists( 'NBPC_Register_Uninstall' ) ) {
 		}
 
 		public function get_items(): Generator {
-			yield null;
+			yield call_user_func( [ NBPC_Registers::class, 'regs_uninstall' ], $this );
 		}
 	}
 }
