+++ core/registers/register-post-type.md
