+++ core/registers/register-taxonomy.md
