+++ core/registers/regs/class-nbpc-reg-activation.php
