+++ core/registers/regs/class-nbpc-reg-ajax.php
