+++ core/registers/regs/class-nbpc-reg-option.php
