+++ core/registers/regs/class-nbpc-reg-post-type.php
