+++ core/registers/regs/class-nbpc-reg-script.php
