+++ core/registers/regs/class-nbpc-reg-shortcode.php
