+++ core/registers/regs/class-nbpc-reg-style.php
