+++ core/registers/regs/class-nbpc-reg-submit.php
