+++ core/registers/regs/class-nbpc-reg-uninstall.php
