+++ core/traits/trait-nbpc-hook-impl.php
