+++ core/traits/trait-nbpc-submodule-impl.php
