+++ core/traits/trait-nbpc-template-impl.php
